homework 7 all editing & adding unavailable online

+++ app/src/main/java/com/example/task/ui/home/HomeFragment.kt
@@ -3,6 +3,7 @@ package com.example.task.ui.home
 import android.app.AlertDialog
 import android.content.DialogInterface
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -12,8 +13,10 @@ import androidx.navigation.fragment.findNavController
 import com.example.task.App
 import com.example.task.R
 import com.example.task.databinding.FragmentHomeBinding
+import com.example.task.utils.isNetworkConnected
 import com.example.task.model.Task
 import com.example.task.ui.home.adapter.TaskAdapter
+import com.example.task.utils.showToast
 
 
 class HomeFragment : Fragment() {
@@ -21,9 +24,8 @@ class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding? = null
     private lateinit var adapter: TaskAdapter
     private lateinit var data: List<Task>
-
-
     private val binding get() = _binding!!
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         adapter = TaskAdapter(this::onLongClick, this::onClick)
@@ -49,33 +51,74 @@ class HomeFragment : Fragment() {
 
     private fun clickListener() {
         binding.fab.setOnClickListener {
-            findNavController().navigate(R.id.navigation_task)
+            if (requireContext().isNetworkConnected()) {//homework7
+                showToast("Adding unavailable now")
+            } else {
+                findNavController().navigate(R.id.navigation_task)
+            }
         }
     }
 
     private fun setData() {
-        data = App.db.dao().getAll()
-        adapter.addTasks(data)
+        if (requireContext().isNetworkConnected()) {
+            /*Get data from firebase*/
+            getData()
+        } else {
+            data = App.db.dao().getAll()
+            adapter.addTasks(data)
+        }
     }
+
+    private fun getData() {
+        App.firebaseDB?.collection("tasks")?.get()?.addOnCompleteListener {
+            if (it.isSuccessful) {
+                val data = arrayListOf<Task>()
+                for (i in it.result) {
+                    val task = i.toObject(Task::class.java)
+                    data.add(task)
+                }
+                adapter.addTasks(data)
+            }
+        }?.addOnFailureListener {
+            Log.e("ololo", "get Data${it.message}")
+        }
+    }
+
+//    private fun deleteData(task: Task) {
+//        App.firebaseDB?.collection("tasks")?.document(data.)?.delete()
+//            ?.addOnSuccessListener { Log.d(TAG, "DocumentSnapshot successfully deleted!") }
+//            ?.addOnFailureListener { Log.e(TAG, "Error deleting document") }
+//    }
 
 
     private fun onLongClick(task: Task) {
         val builder = AlertDialog.Builder(requireContext())
+        if (requireContext().isNetworkConnected()) { //homework7
+            builder.setTitle("Sorry")
+            builder.setMessage("Deleting unavailable now")
+            builder.show()
+        } else {
+            builder.setTitle("Delete")
+            builder.setMessage("Are you sure you want to delete this note?")
 
-        builder.setTitle("Delete")
-        builder.setMessage("Are you sure you want to delete this note?")
 
-        builder.setPositiveButton("Yes") { _: DialogInterface, _: Int ->
-            App.db.dao().delete(task)
-            setData()
+            builder.setPositiveButton("Yes") { _: DialogInterface, _: Int ->
+                App.db.dao().delete(task)
+                setData()
+            }
+
+            builder.setNegativeButton("No") { _: DialogInterface, _: Int ->
+            }
+            builder.show()
         }
-        builder.setNegativeButton("No") { _: DialogInterface, _: Int ->
-        }
-        builder.show()
     }
 
     private fun onClick(task: Task) {
-        findNavController().navigate(R.id.navigation_task, bundleOf(KEY_FOR_TASK to task))
+        if (requireContext().isNetworkConnected()) {//homework7
+            showToast("Editing unavailable now")
+        } else {
+            findNavController().navigate(R.id.navigation_task, bundleOf(KEY_FOR_TASK to task))
+        }
     }
